Added ApiMappable and ApiForm+ObjectMapper
- ApiMappable is a replacement for ApiTransformable protocol. It
extends ObjectMappers Mappable protocol

- Just started on ApiFormMappable. TODO: learn about how swift generics
and protocols work to see if I could implement a generic class that
would work for ApiMappable & ApiTransformable (hopefully get that added
to official ApiModel).
For now just going to create separate classes with Mappable in the name.
// TODO: consider if naming of this file. Not sure if it is correct. It
may change in the future as I get a better idea of how people are
naming extensions in Swift projects. The + feels like an objc
convention?